Predictive Analytics: The Power to Predict Who Will Click, Buy, Lie, or Die
Q & A with Author Eric Siegel
Did Nate Silver use predictive analytics to forecast Obama's 2012 election?
No—but Obama did use predictive analytics to help get elected. Nate Silver made election forecasts for each state as a whole: which way would a state trend, overall? In the meantime, the Obama campaign was using predictive analytics to render per-voter predictions. Moving beyond forecasting, true power comes in influencing the future rather than speculating on it—the raison d'être of predictive analytics. Nate Silver publicly competed to win election forecasting, while Obama's analytics team quietly competed to win the election itself. Specifically, team Obama drove per-voter campaign decisions by way of per-vote predictions.

Why does early retirement predict a shorter life expectancy & why do vegetarians miss fewer flights?
These are two more colorful examples of the multitudes of predictive discoveries waiting within data.

University of Zurich discovered that, for a certain working category of males in Austria, each additional year of early retirement decreases life expectancy by 1.8 months. They conjecture that this could be due to unhealthy habits such as smoking and drinking following retirement.

One airline discovered that customers who preorder a vegetarian meal are more likely to make their flight, with the interpretation that knowledge of a personalized or specific meal awaiting the customer provides an incentive, or establishes a sense of commitment.

Predictive analytics seeks out such predictive connections and then works to see how they may combine together for more precise prediction.

What are the hottest trends in predictive analytics?
There have been many exciting improvements in the core technology of predictive analytics. One is "uplift modeling" (a.k.a. "persuasion modeling"), which predicts influence. ..in order to do influence. The Obama campaign used it to influence voters in the 2012 presidential election; marketing uses it to more adeptly persuade customers; and medicine uses it to better select per-patient treatments. This topic is the focus of the final chapter of this book.

Another hot trend is ensemble models. Like the collective intelligence that spawns the wisdom of a crowd of people, we see the same effect with a crowd of predictive models. Each model alone may be fairly primitive such as a few simple rules, so it gets prediction wrong a lot, as an individual person trying to predict also does. But have them come together as a group and there emerges a new level of predictive performance.

Does the NSA use predictive analytics, and how does that impact the amount of data collected on us?
It's a foregone conclusion that the world's largest spy organization employing the world's largest number of Ph.D. mathematicians considers predictive analytics a strategic priority. Predictive analytics realizes a great potential for law enforcement: The automatic discovery of new suspects. The value of this capability multiplies the incentive to collect increasing amounts of data about civilians. The NSA needs data about everyone, including those of us with no connection to crime whatsoever—not to spy on us but to establish a quantitative baseline. This in turn only amplifies the stakes of the contentious security-versus-privacy debate.

What is the coolest thing predictive analytics has done?
One of the most inspirational accomplishments of predictive analytics is IBM's "Jeopardy!"-playing Watson computer, which triumphed against the all-time human champions on the TV quiz show. The questions can be about most any topic, are intended for humans to answer, and can be complex grammatically. It turns out that predictive modeling is the way in which Watson succeeds in determining the answer to a question: it predicts, "Is this candidate answer the correct answer to this question?" It knocks off one correct answer after another—incredible.

What are companies predicting about me as a customer?
Here are just a few examples:

- Facebook predicts which of 1,500 candidate posts (on average) will be most interesting to you in order to personalize your ordered news feed.

- Microsoft helped develop technology that, based on GPS data, accurately predicts one's location up to multiple years beforehand.

- Target predicts customer pregnancy from shopping behavior, thus identifying prospects to contact with offers related to the needs of a newborn's parents.

- Tesco (UK) annually issues 100 million personalized coupons at grocery cash registers across 13 countries. Predictive analytics increased redemption rates by a factor of 3.6.

- Netflix sponsored a $1 million competition to predict which movies you will like in order to improve movie recommendations.

- One top-five U.S. health insurance company predicts the likelihood an elderly insurance policy holder will die within 18 months in order to trigger end-of-life counseling.